backend/requirements.txt
fastapi
uvicorn
python-multipart
google-genai
python-dotenv
sqlalchemy
pydantic-settings
gTTS